Project-AI_Career_Companion/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-secret-manager==2.18.1
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0
google-auth==2.23.4
google-generativeai==0.3.0
mlflow==2.8.1
python-dotenv==1.0.0
python-multipart==0.0.6

# Additional dependencies for enhanced logging and error handling
python-json-logger>=2.0.7
structlog>=23.2.0  # Optional: For structured logging
sentry-sdk[fastapi]>=1.38.0  # Optional: For error tracking